Move movement system stuff into own dir
