+++ src/movement_gameplay/navigation.rs
@@ -1,6 +1,6 @@
-use crate::movement::CharacterVelocity;
-use crate::navigation::navmesh::read_navmesh;
-use crate::player::Player;
+use crate::movement_gameplay::general_movement::CharacterVelocity;
+use crate::movement_gameplay::navigation::navmesh::read_navmesh;
+use crate::movement_gameplay::player::Player;
 use crate::GameState;
 use bevy::prelude::*;
 use bevy_pathmesh::PathMesh;
