+++ src/saving.rs
@@ -1,6 +1,6 @@
 use crate::condition::ActiveConditions;
 use crate::dialog::{CurrentDialog, DialogEvent};
-use crate::player::Player;
+use crate::movement_gameplay::player::Player;
 use crate::spawning::{DelayedSpawnEvent, GameObject, SpawnEvent};
 use crate::world_serialization::{CurrentLevel, WorldLoadRequest};
 use crate::GameState;
